updates the comments on the time complexity [JS]
COMMENTS:
The time complexity is still quadratic what changes is the string
size that determines it. If the first string has more characters
than the second string, then, the time complexity is determined by
the Brute Force Algorithm that implements the distinct() function.
Otherwise, the time complexity is determined by the linear search
(as before).