adds retro-games App [JS]
COMMENTS:
The Retro Games App stores/fetches info about retro games.

On application startup the App adds an array of games to the
mongoose database (to have something to show).

The App handles GET and POST requests.

The App gets all the games that match a string or simply all the
existing games upon request.

The App adds new games to the database upon POST requests. The
App checks if the game exist in the database to avoid duplication.