removes callback from async operation [JS]
COMMENTS:
Mongoose callbacks should not be used in combination with async/await.

Note that the insertMany() method returns a `Promise' so that we
can chain then() and catch(); however, the exists() method returns
a query (not a Promise) so we use exec() in combination with await
because we have requested asynchronous execution for that method.