revises async/await operations [JS]
COMMENTS:
The keyword `await' is used when waiting asynchronously on methods
that return a Promise.

Not all mongoose methods return a Promise, some return Queries.
When using the latter methods one has to invoke exec() method so
that the method returns a Promise instead.

Moreover, methods that return a Promise are `thenable' and
`catchable'. We have made use of then() and catch() to make the
code more readable.

Adds useful references to the list of references.

App has been tested as in previous commits.