handles games with the same name [JS]
COMMENTS:
The App now informs the users that there is a game with the same
name in the database if they attempt to add a game with the same
name as an existing game to the database.

Note that this was not possible before because the previous
implementation used the mongoose method exists(). Meaning that
games with the same name but different `data' could be inserted to
the database. They had to have identical `data' to be treated as
different.